requirements.txt
blinker==1.9.0
boto3==1.42.74
botocore==1.42.74
click==8.3.1
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.6
